# backend/customer/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Customer API
    version: 0.0.1
paths:
    /customer/estimate-price:
        post:
            tags:
                - Customer
            operationId: Customer_EstimatePrice
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/EstimatePriceRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EstimatePriceResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /customer/get-verify-code/{telephone}:
        get:
            tags:
                - Customer
            operationId: Customer_GetVerifyCode
            parameters:
                - name: telephone
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetVerifyCodeResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /customer/login:
        post:
            tags:
                - Customer
            operationId: Customer_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LoginResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /customer/logout:
        get:
            tags:
                - Customer
            operationId: Customer_Logout
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LogoutResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        EstimatePriceRequest:
            type: object
            properties:
                origin:
                    type: string
                destination:
                    type: string
        EstimatePriceResponse:
            type: object
            properties:
                origin:
                    type: string
                destination:
                    type: string
                price:
                    type: string
                code:
                    type: string
                message:
                    type: string
        GetVerifyCodeResponse:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
                verifyCode:
                    type: string
                verifyCodeTime:
                    type: integer
                    format: int32
                verifyCodeLife:
                    type: integer
                    format: int32
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        LoginRequest:
            type: object
            properties:
                telephone:
                    type: string
                verifyCode:
                    type: string
        LoginResponse:
            type: object
            properties:
                code:
                    type: string
                message:
                    type: string
                token:
                    type: string
                tokenCreateAt:
                    type: string
                tokenLife:
                    type: integer
                    format: int32
        LogoutResponse:
            type: object
            properties:
                code:
                    type: string
                message:
                    type: string
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
tags:
    - name: Customer